requirements.txt
pandas==2.2.2
psycopg2-binary==2.9.9
boto3==1.34.84
python-dotenv==1.0.1